+++ mwcp/utils/stringutils.py
@@ -2,11 +2,19 @@
 Utility used for string conversions.
 """
 
-from future.builtins import str
+import sys
+
+if sys.version_info < (3,):
+    ustr = unicode
+else:
+    ustr = str
 
 
-def convert_to_unicode(input_string):
-    if isinstance(input_string, str):
-        return input_string
+def convert_to_unicode(input_value):
+    if isinstance(input_value, ustr):
+        return input_value
+    elif isinstance(input_value, bytes):
+        return ustr(input_value, encoding='latin1', errors='replace')
     else:
-        return str(input_string, encoding='latin1', errors='replace')
+        return convert_to_unicode(str(input_value))
+